# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Backen Porfolio - OpenAPI 3.0",
    "description": "Especificaciones de los enpoint del Backend",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  },
  "servers": [
    {
      "url": "https://petstore3.swagger.io/api/v3"
    }
  ],
  "tags": [
    {
      "name": "usuarios",
      "description": "Everything about your Pets"
    },
    {
      "name": "solicitudes",
      "description": "Access to Petstore orders"
    }
  ],
  "paths": {
    "/usuarios/createUsuario": {
      "post": {
        "tags": [
          "usuarios"
        ],
        "summary": "Agregar un nuevo usuaruio",
        "description": "Agrega un nuevo usuario",
        "operationId": "addUsuarios",
        "requestBody": {
          "description": "Crear un nuevo usaurio",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuarios"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Usuarios"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Usuarios"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuarios"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Usuarios"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "tags": [
          "usuarios"
        ],
        "summary": "Buscar un usuario por id",
        "description": "Buscar un usuario por id",
        "operationId": "buscarporid",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuarios"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuarios"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/usuarios/login": {
      "get": {
        "tags": [
          "usuarios"
        ],
        "summary": "Loging del usuario",
        "description": "Devuelve token",
        "operationId": "updatePetWithForm",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "Name of pet that needs to be updated",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "Status of pet that needs to be updated",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve el token"
          },
          "405": {
            "description": "Invalid username password supplied"
          }
        }
      }
    },
    "/solicitudes/{usuarioPorfolio}": {
      "get": {
        "tags": [
          "solicitudes"
        ],
        "summary": "Loging del usuario",
        "description": "Devuelve token",
        "operationId": "getsolicitudes",
        "parameters": [
          {
            "name": "usuarioPorfolio",
            "in": "path",
            "description": "ID of pet to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve el token"
          },
          "405": {
            "description": "Invalid username password supplied"
          }
        }
      }
    },
    "/solicitudes/": {
      "post": {
        "tags": [
          "solicitudes"
        ],
        "summary": "Agregar un nueva solicitud",
        "description": "Agregar un nueva solicitud",
        "operationId": "crearSolicitud",
        "parameters": [
          {
            "name": "empresa",
            "in": "query",
            "description": "Name of pet that needs to be updated",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "usuarioPorfolio",
            "in": "query",
            "description": "Status of pet that needs to be updated",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "description": "Status of pet that needs to be updated",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nombre",
            "in": "query",
            "description": "Status of pet that needs to be updated",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apellidos",
            "in": "query",
            "description": "Status of pet that needs to be updated",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mensaje",
            "in": "query",
            "description": "Status of pet that needs to be updated",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fechaAlta",
            "in": "query",
            "description": "Status of pet that needs to be updated",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "telefono",
            "in": "query",
            "description": "Status of pet that needs to be updated",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ususario creado"
          },
          "405": {
            "description": "Invalid"
          }
        }
      }
    },
    "/solicitudes/{id}": {
      "delete": {
        "tags": [
          "solicitudes"
        ],
        "summary": "Elimina una solicitud",
        "description": "Elimina una solicitud",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "eliminio el registro"
          },
          "405": {
            "description": "Invalid"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Usuarios": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Perez"
          },
          "lastname": {
            "type": "string",
            "example": "Juan"
          },
          "email": {
            "type": "string",
            "example": "gaston@gmail.com"
          },
          "password": {
            "type": "string",
            "format": "1234567"
          }
        }
      },
      "Solicitudes": {
        "type": "object",
        "properties": {
          "empresa": {
            "type": "string",
            "example": "UADE"
          },
          "usuarioPorfolio": {
            "type": "string",
            "example": "gaston@gmail.com"
          },
          "nombre": {
            "type": "string",
            "example": "Gaston"
          },
          "apellidos": {
            "type": "string",
            "example": "Gomez"
          },
          "email": {
            "type": "string",
            "example": "gastongmz@gmail.com"
          },
          "telefono": {
            "type": "string",
            "example": "45563333"
          },
          "mensaje": {
            "type": "string",
            "example": "cantacar por favor"
          },
          "fechaAlta": {
            "type": "string",
            "example": "1-1-2023"
          }
        }
      }
    },
    "requestBodies": {
      "wfasdf": {
        "description": "Pet object that needs to be added to the store",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Solicitudes"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Solicitudes"
            }
          }
        }
      },
      "UserArray": {
        "description": "List of user object",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Usuarios"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "solicitudesstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://petstore3.swagger.io/oauth/authorize",
            "scopes": {
              "write:usuarios": "modify pets in your account",
              "read:usuarios": "read your pets"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}